Refactor bot.py into respective function files
