+++ .gitignore
@@ -2,3 +2,4 @@ data_file.json
 __pycache__
 .DS_Store
 env/
+.vscode/
